requirements.txt
pandas==2.2.0
numpy==1.26.3
pyyaml==6.0.1
pydantic==2.6.0
python-dotenv==1.0.1
jinja2==3.1.3
markupsafe==2.1.5